requirements.txt
# Core OCR and PDF processing
pdf2image>=3.1.0
pytesseract>=0.3.10
Pillow>=9.0.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0


